seperate function for devices tcp
