+++ src/tcp/peripherals_tcp_parser.rs
@@ -1,13 +1,8 @@
 use crate::ble_device_handlers;
 use crate::ble_device_handlers::smart_bike_trainer;
-use anyhow::Context;
-use anyhow::Result;
-use anyhow::anyhow;
 
-use crate::{
-    ble_device_handlers::{self, BleDevice},
-    tcp, tcp_parser,
-};
+use crate::tcp::device_tcp_parser;
+use crate::{ble_device_handlers::BleDevice, tcp, tcp_parser};
 use anyhow::{Context, Result, anyhow};
 use btleplug::{
     api::{BDAddr, Peripheral},
@@ -18,20 +13,6 @@ use spdlog::prelude::*;
 use std::{collections::HashSet, time::Duration};
 use tokio::{net::TcpStream, time::sleep};
 
-use crate::{ble_device_handlers::BleDevice, tcp};
-
-const SMART_TRAINER_DEVICE_TYPE: &str = "smart trainer";
-pub async fn send_device_connection_information(
-    stream: &mut TcpStream,
-    index: usize,
-    device_type_name: &str,
-) {
-    tcp::send_tcp_data(stream, format!("o|{}|[{}]", device_type_name, index)).await;
-}
-pub async fn send_bike_trainer_data(stream: &mut TcpStream, power: u16, cadence: u16) {
-    tcp::send_tcp_data(stream, format!("p{}", power)).await;
-    tcp::send_tcp_data(stream, format!("c{}", cadence)).await;
-}
 pub async fn send_peripherals(
     stream: &mut TcpStream,
     peripherals: &[btleplug::platform::Peripheral],
@@ -75,6 +56,7 @@ pub async fn send_peripherals(
     }
 }
 
+const SMART_TRAINER_DEVICE_TYPE: &str = "smart trainer";
 pub(super) async fn handle_parsing_peripheral_connection(
     data: &str,
     valid_peripherals: &[btleplug::platform::Peripheral],
@@ -121,6 +103,7 @@ pub(super) async fn handle_parsing_peripheral_connection(
     }
     // only if connection to device succeed
     info!("Successfully connected to device! sending connection information thru tcp");
-    tcp_parser::send_device_connection_information(stream, device_index, device_type_name).await;
+    device_tcp_parser::send_device_connection_information(stream, device_index, device_type_name)
+        .await;
     Ok(Some(device))
 }
